+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 ~~~~~~~
 
+1.7.2 (unreleased)
+------------------
+
+
 1.7.1 (2013-09-07)
 ------------------
 - launchpad #1215838: don't freeze develop / gp.vcsdevelop distributions